style: Crear responsive para Header y NavBar
